+++ MapboxSearch/place-autocomplete/src/main/java/com/mapbox/search/autocomplete/PlaceAutocomplete.kt
@@ -57,7 +57,7 @@ public interface PlaceAutocomplete {
     public suspend fun select(suggestion: PlaceAutocompleteSuggestion): Expected<Exception, PlaceAutocompleteResult>
 
     /**
-     * @suppress
+     * Companion object.
      */
     public companion object {