+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/DataClassGeneratedFunctionsTest.kt
@@ -5,6 +5,7 @@ import com.mapbox.search.analytics.FeedbackEvent
 import com.mapbox.search.analytics.MissingResultFeedbackEvent
 import com.mapbox.search.common.IsoCountryCode
 import com.mapbox.search.common.IsoLanguageCode
+import com.mapbox.search.common.NavigationProfile
 import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.common.metadata.ImageInfo
 import com.mapbox.search.common.metadata.OpenHours
@@ -40,7 +41,7 @@ internal class DataClassGeneratedFunctionsTest {
         listOf(
             // regular classes and objects
             SearchOptions::class,
-            SearchNavigationProfile::class,
+            NavigationProfile::class,
             SearchNavigationOptions::class,
             RouteOptions::class,
             RouteOptions.Deviation.SarType::class,